backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pydantic>=2.7.0
aiosqlite>=0.20.0
